requirements.txt
annotated-types==0.7.0
anykeystore==0.2
apex==0.9.10.dev0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.8
contourpy==1.1.0
cryptacular==1.6.2
cycler==0.11.0
decord==0.6.0
defusedxml==0.7.1
docker-pycreds==0.4.0
dotmap==1.3.30
einops==0.6.1
eval_type_backport==0.2.2
filelock==3.12.4
fonttools==4.40.0
fsspec==2023.9.2
ftfy==6.1.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==2.0.2
huggingface-hub==0.17.2
hupper==1.12
idna==3.4
importlib-resources==5.12.0
joblib==1.3.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
numpy==1.20.3
nvidia-ml-py==11.525.112
nvitop==1.1.2
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
PasteDeploy==3.0.1
pbkdf2==1.3
Pillow==9.5.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.3.6
protobuf==5.29.3
psutil==5.9.5
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.0.9
pyramid==2.0.1
pyramid-mailer==0.15.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==5.4.1
randaugment==1.0.2
regex==2023.6.3
repoze.sendmail==4.4.1
requests==2.31.0
requests-oauthlib==1.3.1
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==2.22.0
setproctitle==1.3.5
six==1.16.0
smmap==5.0.2
SQLAlchemy==2.0.16
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==1.13.0+cu116
torchaudio==0.13.0+cu116
torchvision==0.14.0+cu116
tqdm==4.65.0
transaction==3.1.0
transformers==4.33.2
translationstring==1.4
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.0.3
velruse==1.1.1
venusian==3.0.0
wandb==0.19.7
wcwidth==0.2.6
WebOb==1.8.7
WTForms==3.0.1
wtforms-recaptcha==0.3.2
yacs==0.1.8
zipp==3.15.0
zope.deprecation==5.0
zope.interface==6.0
zope.sqlalchemy==3.0
